[refactor] Separate stories into files
